chore: dependency update
this updates some core dependencies to their latest version and fixes the tests to be able to fully run again.

Fixing the tests turned out to be a rather large operation, since the existing tests did rely on implementation details of `@googlemaps/js-api-loader` that have been changed in newer versions. This is now replaced with a proper mock-implementation.